WORLDSERVICE-543: Rename file.
